Separate BaseSonarrCmdlet
